Implement tooltip with participants in connected profiles component
