+++ webapp/src/components/channel_link_label/component.tsx
@@ -6,7 +6,7 @@ import {OverlayTrigger, Tooltip} from 'react-bootstrap';
 import {Channel} from 'mattermost-redux/types/channels';
 import {UserProfile} from 'mattermost-redux/types/users';
 
-import {getUserDisplayName} from '../../utils';
+import {getUserDisplayName, getUsersList} from '../../utils';
 
 import ActiveCallIcon from '../../components/icons/active_call_icon';
 
@@ -15,19 +15,6 @@ interface Props {
     hasCall: boolean,
     profiles: UserProfile[],
 }
-
-const getUsersList = (profiles: UserProfile[]) => {
-    if (profiles.length === 0) {
-        return '';
-    }
-    if (profiles.length === 1) {
-        return getUserDisplayName(profiles[0]);
-    }
-    const list = profiles.slice(0, -1).map((profile, idx) => {
-        return getUserDisplayName(profile);
-    }).join(', ');
-    return list + ' and ' + getUserDisplayName(profiles[profiles.length - 1]);
-};
 
 const ChannelLinkLabel = (props: Props) => {
     if (props.hasCall) {
